fix(extension): keep search controls in graph view
