+++ .changeset/show-workspace-graph-title.md
@@ -2,6 +2,6 @@
 "@codegraphy-dev/extension": minor
 ---
 
-Show the current workspace name as the Graph view title and keep search controls in a compact Search view above it.
+Show the current workspace name as the Graph view title and keep search controls inside the graph.
 
-The Graph view now shows the current workspace name in VS Code's view title, falling back to Graph when no workspace is open. Search, match-case/whole-word/regex toggles, and filter controls live in their own compact Search view above the graph and stay available while graph data is loading or refreshing.
+The Graph view now shows the current workspace name in VS Code's view title, falling back to Graph when no workspace is open. Search, match-case/whole-word/regex toggles, and filter controls remain in the Graph view header and stay available while graph data is loading or refreshing.
