+++ packages/extension/src/extension/activate.ts
@@ -45,22 +45,8 @@ export function activate(context: vscode.ExtensionContext): CodeGraphyAPI {
     context: { workspaceFolders: vscode.workspace.workspaceFolders?.length ?? 0 },
   });
   const provider = new GraphViewProvider(context.extensionUri, context);
-  void vscode.commands.executeCommand(
-    'setContext',
-    GraphViewProvider.searchControlsVisibleContext,
-    false,
-  );
 
   context.subscriptions.push(
-    vscode.window.registerWebviewViewProvider(
-      GraphViewProvider.searchViewType,
-      provider,
-      {
-        webviewOptions: {
-          retainContextWhenHidden: true,
-        },
-      }
-    ),
     vscode.window.registerWebviewViewProvider(
       GraphViewProvider.viewType,
       provider,
@@ -90,7 +76,7 @@ export function activate(context: vscode.ExtensionContext): CodeGraphyAPI {
   diagnostics.emit({
     area: 'extension.lifecycle',
     event: 'activation-completed',
-    context: { registeredWebviewProviders: 3 },
+    context: { registeredWebviewProviders: 2 },
   });
 
   return {
