+++ packages/extension/src/webview/app/shell/view.tsx
@@ -1,7 +1,9 @@
 import React, { useEffect, useState } from 'react';
+import type { SearchOptions } from '../../components/searchBar/field/model';
 import { useTheme } from '../../theme/useTheme';
 import { usePluginManager } from '../../pluginRuntime/useManager';
 import { useFilteredGraph } from '../../search/useFilteredGraph';
+import { postMessage } from '../../vscodeApi';
 import { getNoDataHint } from './messages';
 import { setupMessageListener } from './messageListener';
 import { LoadingState, EmptyState } from './states';
@@ -15,9 +17,11 @@ import { PanelStack } from './panel/stack';
 import { ToolbarRail } from './panel/toolbar';
 import { useFilterLegendInputs } from './derivedState';
 import { useRulePromptHandlers } from '../rulePrompt/handlers';
-import { buildPendingFilterPatterns } from './filterPopover';
+import { buildPendingFilterPatterns, useFilterPopoverState } from './filterPopover';
 import { useVisibleGraphStateResponse } from './visibleGraphResponse';
-import { ActiveFileBreadcrumb } from '../../components/activeFileBreadcrumb/view';
+import { SearchHeader } from './panel/search';
+import { useShellVisibleGraphs } from './visibleGraphs';
+import { getShellGraphCountState } from './counts';
 
 export default function App(): React.ReactElement {
   const { pluginHost, injectPluginAssets, resetPluginAssets } = usePluginManager();
@@ -30,6 +34,7 @@ export default function App(): React.ReactElement {
     legends,
     filterPatterns,
     pluginFilterPatterns,
+    pluginFilterGroups,
     disabledCustomFilterPatterns,
     disabledPluginFilterPatterns,
     showOrphans,
@@ -49,7 +54,11 @@ export default function App(): React.ReactElement {
   } = useAppState();
   const {
     setActivePanel,
+    setSearchQuery,
+    setSearchOptions,
     setFilterPatterns,
+    setDisabledCustomFilterPatterns,
+    setDisabledPluginFilterPatterns,
   } = useAppActions();
   const setOptimisticUserLegends = useGraphStore((state) => state.setOptimisticUserLegends);
   const [rulePrompt, setRulePrompt] = useState<RulePromptState | null>(null);
@@ -64,8 +73,15 @@ export default function App(): React.ReactElement {
   );
   const effectiveShowOrphans = graphHasIndex ? showOrphans : true;
   const {
+    filterPopoverOpen,
+    handleFilterPopoverOpenChange,
+    pendingFilterPatterns,
+  } = useFilterPopoverState();
+  const {
     coloredData,
+    filteredData,
     edgeDecorations: graphEdgeDecorations,
+    regexError,
   } = useFilteredGraph(
     graphData,
     searchQuery,
@@ -100,6 +116,30 @@ export default function App(): React.ReactElement {
 
   const displayGraphData = coloredData || graphData;
   useVisibleGraphStateResponse(displayGraphData);
+  const { countBaseData, filterVisibleData } = useShellVisibleGraphs({
+    activeFilterPatterns,
+    edgeVisibility,
+    graphData,
+    graphEdgeTypes,
+    graphNodeTypes,
+    nodeVisibility,
+    searchOptions,
+    showOrphans: effectiveShowOrphans,
+  });
+  const { countState, countTotal, excludedCount } = graphData
+    ? getShellGraphCountState({
+      countBaseData,
+      filterVisibleData,
+      filteredData,
+      graphData,
+      regexError,
+      searchQuery,
+    })
+    : {
+      countState: { label: null },
+      countTotal: 0,
+      excludedCount: 0,
+    };
 
   const closeActivePanel = () => setActivePanel('none');
   const openFilterPromptWithPatterns = (patterns: string[]) => {
@@ -107,6 +147,20 @@ export default function App(): React.ReactElement {
     if (pattern) {
       openFilterPrompt(pattern);
     }
+  };
+  const commitSearchState = (query: string, options: SearchOptions): void => {
+    postMessage({
+      type: 'UPDATE_SEARCH_STATE',
+      payload: { query, options },
+    });
+  };
+  const handleSearchQueryChange = (query: string): void => {
+    setSearchQuery(query);
+    commitSearchState(query, searchOptions);
+  };
+  const handleSearchOptionsChange = (options: SearchOptions): void => {
+    setSearchOptions(options);
+    commitSearchState(searchQuery, options);
   };
   const graphContent = (() => {
     if (isLoading) {
@@ -158,11 +212,31 @@ export default function App(): React.ReactElement {
 
   return (
     <div className="relative w-full h-screen flex flex-col">
-      {activeFilePath && (
-        <div className="flex-shrink-0 border-b border-border px-2 py-1">
-          <ActiveFileBreadcrumb filePath={activeFilePath} />
-        </div>
-      )}
+      <SearchHeader
+        searchQuery={searchQuery}
+        searchOptions={searchOptions}
+        resultCount={filteredData?.nodes.length}
+        totalCount={countTotal}
+        activeFilePath={activeFilePath}
+        countLabel={countState.label}
+        filterPopover={{
+          customPatterns: filterPatterns,
+          disabledCustomPatterns: disabledCustomFilterPatterns,
+          disabledPluginPatterns: disabledPluginFilterPatterns,
+          excludedCount,
+          onDisabledCustomPatternsChange: setDisabledCustomFilterPatterns,
+          onDisabledPluginPatternsChange: setDisabledPluginFilterPatterns,
+          onOpenChange: handleFilterPopoverOpenChange,
+          onPatternsChange: setFilterPatterns,
+          open: filterPopoverOpen,
+          pendingPatterns: pendingFilterPatterns,
+          pluginGroups: pluginFilterGroups,
+          pluginPatterns: pluginFilterPatterns,
+        }}
+        regexError={regexError}
+        onSearchQueryChange={handleSearchQueryChange}
+        onSearchOptionsChange={handleSearchOptionsChange}
+      />
       <div className="flex-1 min-h-0 relative">
         {graphContent}
       </div>
