+++ packages/extension/tests/extension/extension.test.ts
@@ -32,11 +32,6 @@ describe('Extension', () => {
       activate(mockContext as unknown as Parameters<typeof activate>[0]);
 
       expect(vscode.window.registerWebviewViewProvider).toHaveBeenCalledWith(
-        'codegraphy.searchControlsView',
-        expect.any(Object),
-        { webviewOptions: { retainContextWhenHidden: true } }
-      );
-      expect(vscode.window.registerWebviewViewProvider).toHaveBeenCalledWith(
         'codegraphy.graphView',
         expect.any(Object),
         { webviewOptions: { retainContextWhenHidden: true } }
@@ -45,16 +40,6 @@ describe('Extension', () => {
         'codegraphy.timelineView',
         expect.any(Object),
         { webviewOptions: { retainContextWhenHidden: true } }
-      );
-    });
-
-    it('hides the search controls until the graph view is ready', () => {
-      activate(mockContext as unknown as Parameters<typeof activate>[0]);
-
-      expect(vscode.commands.executeCommand).toHaveBeenCalledWith(
-        'setContext',
-        'codegraphy.searchControlsVisible',
-        false
       );
     });
 
@@ -79,11 +64,11 @@ describe('Extension', () => {
     it('should add subscriptions to context', () => {
       activate(mockContext as unknown as Parameters<typeof activate>[0]);
 
-      // view providers (3) + config listener (1) + active editor listener (1) + save listener (1)
+      // view providers (2) + config listener (1) + active editor listener (1) + save listener (1)
       // + file watcher events (2) + workspace file operation events (3) + file watcher (1)
       // + URI handler (1) + runtime bridge listener (1)
       // + 16 commands (open, openInEditor, fitView, zoomIn, zoomOut, undo, redo, exportPng, exportSvg, exportJpeg, exportJson, exportMarkdown, clearCache, toggleDepthMode, cycleLayout, toggleDimension)
-      expect(mockContext.subscriptions.length).toBe(30);
+      expect(mockContext.subscriptions.length).toBe(29);
     });
 
     it('should ignore workspace settings saves when deciding graph refresh', async () => {
