+++ packages/extension/tests/extension/extensionIndex.test.ts
@@ -53,11 +53,6 @@ describe('extension activate API', () => {
     activate(mockContext as never);
 
     expect(vscode.window.registerWebviewViewProvider).toHaveBeenCalledWith(
-      'codegraphy.searchControlsView',
-      expect.any(Object),
-      { webviewOptions: { retainContextWhenHidden: true } }
-    );
-    expect(vscode.window.registerWebviewViewProvider).toHaveBeenCalledWith(
       'codegraphy.graphView',
       expect.any(Object),
       { webviewOptions: { retainContextWhenHidden: true } }
