+++ packages/extension/tests/extension/graphView/provider/webview/resolve.test.ts
@@ -3,84 +3,6 @@ import * as vscode from 'vscode';
 import { resolveGraphViewProviderWebviewView } from '../../../../../src/extension/graphView/provider/webview/resolve';
 
 describe('graphView/provider/webview/resolve', () => {
-  it('reveals the search controls after the visible graph view resolves', async () => {
-    vi.useFakeTimers();
-    const webviewView = {
-      viewType: 'codegraphy.graphView',
-      webview: {},
-      visible: true,
-      onDidChangeVisibility: vi.fn(() => ({ dispose: vi.fn() })),
-      onDidDispose: vi.fn(() => ({ dispose: vi.fn() })),
-    } as unknown as vscode.WebviewView;
-    const executeCommand = vi.fn(() => Promise.resolve(undefined));
-    const source = {
-      _extensionUri: vscode.Uri.file('/test/extension'),
-      _view: undefined,
-      _timelineView: undefined,
-      _getLocalResourceRoots: vi.fn(() => []),
-      flushPendingWorkspaceRefresh: vi.fn(),
-    };
-
-    resolveGraphViewProviderWebviewView(source as never, {
-      createHtml: vi.fn(() => '<graph html />'),
-      executeCommand,
-      getWorkspaceTitle: vi.fn(() => 'CodeGraphyV4'),
-      resolveWebviewView: vi.fn(),
-      setWebviewMessageListener: vi.fn(),
-    }, webviewView);
-
-    expect(executeCommand).not.toHaveBeenCalledWith(
-      'setContext',
-      'codegraphy.searchControlsVisible',
-      true
-    );
-
-    await vi.runOnlyPendingTimersAsync();
-
-    expect(executeCommand).toHaveBeenCalledWith(
-      'setContext',
-      'codegraphy.searchControlsVisible',
-      true
-    );
-    vi.useRealTimers();
-  });
-
-  it('does not reveal the search controls when the search view resolves', async () => {
-    vi.useFakeTimers();
-    const webviewView = {
-      viewType: 'codegraphy.searchControlsView',
-      webview: {},
-      visible: true,
-      onDidChangeVisibility: vi.fn(() => ({ dispose: vi.fn() })),
-      onDidDispose: vi.fn(() => ({ dispose: vi.fn() })),
-    } as unknown as vscode.WebviewView;
-    const executeCommand = vi.fn(() => Promise.resolve(undefined));
-    const source = {
-      _extensionUri: vscode.Uri.file('/test/extension'),
-      _searchView: undefined,
-      _view: undefined,
-      _timelineView: undefined,
-      _getLocalResourceRoots: vi.fn(() => []),
-      flushPendingWorkspaceRefresh: vi.fn(),
-    };
-
-    resolveGraphViewProviderWebviewView(source as never, {
-      createHtml: vi.fn(() => '<search html />'),
-      executeCommand,
-      resolveWebviewView: vi.fn(),
-      setWebviewMessageListener: vi.fn(),
-    }, webviewView);
-
-    await vi.runOnlyPendingTimersAsync();
-
-    expect(executeCommand).not.toHaveBeenCalledWith(
-      'setContext',
-      'codegraphy.searchControlsVisible',
-      true
-    );
-    vi.useRealTimers();
-  });
-
   it('titles the graph view with the current workspace name', () => {
     const webviewView = {
       viewType: 'codegraphy.graphView',
@@ -274,55 +196,6 @@ describe('graphView/provider/webview/resolve', () => {
 
     expect(source._timelineView).toBeUndefined();
     expect(source._view).toBe(webviewView);
-  });
-
-  it('resolves the search view independently without resizing the side bar', () => {
-    let disposeListener: (() => void) | undefined;
-    const webview = {
-      options: {},
-      html: '',
-    } as unknown as vscode.Webview;
-    const webviewView = {
-      viewType: 'codegraphy.searchControlsView',
-      webview,
-      visible: true,
-      onDidChangeVisibility: vi.fn(() => undefined),
-      onDidDispose: vi.fn(listener => {
-        disposeListener = listener;
-        return { dispose: vi.fn() };
-      }),
-    } as unknown as vscode.WebviewView;
-    const graphView = { viewType: 'codegraphy.graphView' } as unknown as vscode.WebviewView;
-    const createHtml = vi.fn(() => '<search html />');
-    const executeCommand = vi.fn(() => Promise.resolve(undefined));
-    const resolveWebviewView = vi.fn((_view, options) => {
-      options.getHtml(webview);
-    });
-    const source = {
-      _extensionUri: vscode.Uri.file('/test/extension'),
-      _searchView: undefined,
-      _view: graphView,
-      _timelineView: undefined,
-      _getLocalResourceRoots: vi.fn(() => []),
-      flushPendingWorkspaceRefresh: vi.fn(),
-    };
-
-    resolveGraphViewProviderWebviewView(source as never, {
-      createHtml,
-      executeCommand,
-      resolveWebviewView,
-      setWebviewMessageListener: vi.fn(),
-    }, webviewView);
-
-    expect(source._searchView).toBe(webviewView);
-    expect(source._view).toBe(graphView);
-    expect(createHtml).toHaveBeenCalledWith(source._extensionUri, webview, 'search');
-    expect(source.flushPendingWorkspaceRefresh).not.toHaveBeenCalled();
-    expect(executeCommand).not.toHaveBeenCalled();
-
-    disposeListener?.();
-
-    expect(source._searchView).toBeUndefined();
   });
 
   it('keeps a different graph view attached when the timeline view disposes', () => {
