+++ packages/extension/tests/webview/app/shell/view.test.tsx
@@ -461,10 +461,10 @@ describe('App', () => {
     expect(screen.getByText('2 nodes • 1 connection')).toBeInTheDocument();
   });
 
-  it('does not render search controls inside the graph view while graph data loads', async () => {
+  it('keeps search controls inside the graph view while graph data loads', async () => {
     render(<App />);
 
-    expect(screen.queryByPlaceholderText('Search files... (Ctrl+F)')).not.toBeInTheDocument();
+    expect(screen.getByPlaceholderText('Search files... (Ctrl+F)')).toBeInTheDocument();
     expect(screen.getByText('Loading graph...')).toBeInTheDocument();
 
     await act(async () => {
@@ -478,7 +478,7 @@ describe('App', () => {
       sendMessage({ type: 'APP_BOOTSTRAP_COMPLETE' });
     });
 
-    expect(screen.queryByPlaceholderText('Search files... (Ctrl+F)')).not.toBeInTheDocument();
+    expect(screen.getByPlaceholderText('Search files... (Ctrl+F)')).toBeInTheDocument();
     expect(screen.queryByText('Loading graph...')).not.toBeInTheDocument();
   });
 
@@ -494,7 +494,7 @@ describe('App', () => {
 
     render(<App />);
 
-    expect(screen.queryByPlaceholderText('Search files... (Ctrl+F)')).not.toBeInTheDocument();
+    expect(screen.getByPlaceholderText('Search files... (Ctrl+F)')).toBeInTheDocument();
     expect(screen.getByRole('button', { name: 'Open src/App.ts' })).toBeInTheDocument();
   });
